requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-binance==0.5.13
requests==2.18.4
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
zope.interface==4.4.3
